pgAdmin Table created
Adding the query to create the table. Had some errors at the last line of original code. Dropped "FOREIGN KEY (primary_key) REFERENCES winemag-data_first150k(_c0)
to avoid the error, but may need to fix that if that information is needed.
Added a png of the table headers that were created.